pyproject.toml
[project]
name = "mdsist"
version = "0.0.1"
description = "Machine Learning project around MNIST dataset"
authors = [
  { name = "MDSIST Team" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    
]
requires-python = "~=3.9"

[tool.black]
line-length = 99
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
)/
'''

[tool.ruff.lint.isort]
known_first_party = ["mdsist"]
force_sort_within_sections = true

[tool.poetry]
name = "mdsist"
version = "0.1.0"
description = "Machine learning project around MNIST dataset"
authors = ["pol-arevalo-soler <pol.arevalo@estudiantat.upc.edu>, adriaag <adria.aumatell@estudiantat.upc.edu>, rubenvb01 <ruben.villanueva@estudiantat.upc.edu>,  jizyong@gmail.com <zhengyong.ji@estudiantat.upc.edu>, gnscc <ignasi.cervero@estudiantat.upc.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.1.4"
numpy = "^1.26.4"
scikit-learn = "^1.3.2"
pyarrow = "^17.0.0"
mlflow = "^2.16.2"
typer = "^0.12.5"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
torch = [
  { version = "<=2.2.2", markers = "sys_platform == 'darwin'", source="pypi"},
  { version = "2.4.1", markers = "sys_platform == 'linux'", source="torch"}
]
torchvision = [
  { version = "<=0.17.2", markers = "sys_platform == 'darwin'", source="pypi"},
  { version = "0.19.1", markers = "sys_platform == 'linux'", source="torch"}
]
codecarbon = "^2.7.1"
ptflops = "^0.7.4"
torch-summary = "^1.4.5"
deepchecks = {extras = ["vision"], version = "^0.18.1"}
fastapi = {extras = ["standard"], version = "^0.115.4"}
uvicorn = "^0.32.0"
torchinfo = "^1.8.0"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.1"
pytest = "^8.3.3"
isort = "^5.13.2"
pre-commit = "^3.8.0"
ipykernel = "^6.29.5"
dvc = "^3.55.2"
ptflops = "^0.7.4"
black = {extras = ["jupyter"], version = "^24.8.0"}
pynblint = "^0.1.6"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"